requirements.txt
autopep8==1.3.2
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
pycodestyle==2.3.1
pytz==2017.2
